Rename desktop->wide mobile->narrow
