+++ src/lib/components/navbar/Navlink.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-	import DesktopNavlink from './desktop/DesktopNavlink.svelte'
-	import MobileNavlink from './mobile/MobileNavlink.svelte'
+	import WideNavlink from './wide/WideNavlink.svelte'
+	import NarrowNavlink from './narrow/NarrowNavlink.svelte'
 
 	export let href: string | undefined = undefined
 	export let c2a = false
@@ -12,23 +12,23 @@
 	export let external = false
 </script>
 
-<span class="desktop-navlink">
-	<DesktopNavlink {href} {c2a} {ariaLabel} {inverted} {first} {narrow} {active} {external}>
+<span class="wide-navlink">
+	<WideNavlink {href} {c2a} {ariaLabel} {inverted} {first} {narrow} {active} {external}>
 		<slot />
-	</DesktopNavlink>
+	</WideNavlink>
 </span>
-<span class="mobile-navlink">
-	<MobileNavlink {href} {c2a} {ariaLabel} {inverted} {first} {narrow} {active} {external}>
+<span class="narrow-navlink">
+	<NarrowNavlink {href} {c2a} {ariaLabel} {inverted} {first} {narrow} {active} {external}>
 		<slot />
-	</MobileNavlink>
+	</NarrowNavlink>
 </span>
 
 <style>
-	:global(.mobile-navbar) .desktop-navlink {
+	:global(.narrow-navbar) .wide-navlink {
 		display: none;
 	}
 
-	:global(.desktop-navbar) .mobile-navlink {
+	:global(.wide-navbar) .narrow-navlink {
 		display: none;
 	}
 </style>
