+++ src/lib/components/navbar/narrow/NarrowNavbar.svelte
