+++ src/lib/components/navbar/wide/WideNavbar.svelte
