+++ src/lib/components/navbar/wide/WideNavlink.svelte
